refactor: 프로젝트 구조 개선을 위한 디렉토리 계층 추가
- domain과 global 상위 디렉토리 추가하여 코드 구조화
