+++ src/main/java/com/market/saessag/domain/product/repository/ProductRepository.java
@@ -1,6 +1,6 @@
-package com.market.saessag.product.repository;
+package com.market.saessag.domain.product.repository;
 
-import com.market.saessag.product.entity.Product;
+import com.market.saessag.domain.product.entity.Product;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
