+++ src/main/java/com/market/saessag/domain/user/repository/UserRepository.java
@@ -4,8 +4,7 @@ import com.market.saessag.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
-
-    //  User findByEmailAndPassword(String email, String password);
+    User findByEmail(String email);
     boolean existsByEmail(String email);
     /*
         existsByEmail(): 이메일 중복 검증 방법
