+++ src/test/java/com/market/saessag/product/ProductGenerateData.java
@@ -1,9 +1,9 @@
 package com.market.saessag.product;
 
-import com.market.saessag.product.entity.Product;
-import com.market.saessag.product.repository.ProductRepository;
-import com.market.saessag.user.entity.User;
-import com.market.saessag.user.repository.UserRepository;
+import com.market.saessag.domain.product.entity.Product;
+import com.market.saessag.domain.product.repository.ProductRepository;
+import com.market.saessag.domain.user.entity.User;
+import com.market.saessag.domain.user.repository.UserRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
